+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "C_Cpp.intelliSenseEngineFallback": "disabled",
     "files.associations": {
-        "ostream": "cpp"
+        "ostream": "cpp",
+        "string": "cpp"
     }
 }
